FastSmrMapsLoader without support of checkpoints/trim
First version of the FastSmrMapsLoader that will reconstruct all the SMR
Maps present in the log in one pass. It reads sequentially all entries
from 0 to log tail and dispatch the SMREntries to its map backed by a
Corfu stream. Every time it encounters a new stream, it creates a new Corfu
Object.